[celery] use module import hooks
