+++ tests/contrib/celery/test_patch.py
@@ -2,8 +2,21 @@ import unittest
 from nose.tools import ok_
 from ddtrace import Pin
 
+from tests.contrib import PatchMixin
 
-class CeleryPatchTest(unittest.TestCase):
+
+class CeleryPatchTest(unittest.TestCase, PatchMixin):
+    def test_patch(self):
+        from ddtrace import patch
+        patch(celery=True)
+        import celery
+
+        app = celery.Celery()
+        ok_(Pin.get_from(app) is not None)
+
+    def test_patch_idempotent(self):
+        pass
+
     def test_patch_after_import(self):
         import celery
         from ddtrace import patch
@@ -19,4 +32,3 @@ class CeleryPatchTest(unittest.TestCase):
 
         app = celery.Celery()
         ok_(Pin.get_from(app) is not None)
-
